+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Features:
     * resume: resumes audio object.
     * isplaying: returns if audio object is playing.
     * audioread: read an audio file.
+    * audioinfo: get information about audio file.
 
   * JSON module: 
     * jsonencode: encodes a Nelson object into a JSON string.